Use function for AppSignal alias + function
I keep entering "a server" if I want to switch to the "server"
repository. Instead I should do "a; cd server", but I keep forgetting to
press enter after "a".

So because I'm lazy, let's make this alias a function so I can type "a
server" and end up where I expect.