Fix MacVim focus across spaces
When in a space without the MacVim app the "find first window and focus
that" method didn't work as the `application:allWindows()` method only
returns windows that are present on that space. Triggering MacVim would
do nothing, not even raise an error.

Calling `application:activate()` seems to focus the screen and switch
space if necessary.