Configure vim path so `gf` works
When using `gf` on a `require "filename"` statement it will open the
file in vim. This doesn't work without the `path` being set in vim.

Add more entries to `wildignore` so that these directories don't have
to be scanned fully on every autocomplete request. They usually don't
include anything relevant for the autocomplete, as I want it to be
project code specific.