Auto switch to ergodox keyboard layout on connect
Listen to the event when the ErgoDox EZ keyboard gets connected and
automatically switch the keyboard layout to the ErgoDox layout.

This prevents me from switching manually all the time.

Also check if the ErgoDox is connected on hammerspoon load, also on
configuration reload, so that I'm immediately in the right keyboard
layout and don't have to manually switch.

I don't have to switch keyboard layouts a lot nowadays, but when I'm
tinkering with my hammerspoon config like now I constantly had to switch
keyboard layouts, now it's done automatically on (re)load.