Fix Todoist app keyboard shortcut
The meh + M combination couldn't be typed with my Ergodox keyboard
layout, so update it to use a key combination that is still available.
I'm reaching for new combinations and this isn't ideal, but the easiest
solution for now. Instead I need to remap some of the "management" layer
keys to use the period and comma keys instead.