Switch to wireless headphones shortcut
Update the existing audio device shortcut to switch between speakers and
wireless headphones.

Because the wireless headphones may not be on or connect at times, add
error handling for when this happens that doesn't crash everything.

I also had to update the look up to `findOutputByName` instead of
`findDeviceByName`, because I think it also picks up the microphone of
the headphones as an audio device, which can't be made the output
device.