Remove custom color config from ZSH config
I added this way back in 2015 to match the tig bars with my vim theme in
commit 4c4e8f3a7044b1ed98045c6b8e6852936c4a988b. The commit doesn't list
any sources why or how this works. I looked at the docs and it seems
like a lot of work and config for such as small thing.

I've removed it in favor of using one of the existing 256 colors. The
theme doesn't match anymore, but that's not that important.

These additional prints at the start of the ZSH shell ended up in some
output of commands I want to use,
e.g. `chruby-exec ruby-3.1 -- echo 123 | pbcopy`
It's too much hassle to work around it, so remove it.