Improve keyboard auto switcher
Fix issues with the keyboard auto switcher when I (dis)connect the
ErgoDox. Not always does the USB event get picked up. Checking also what
devices are connected when the laptop wakes or gets unlocked seems more
reliable.

This requires less trial-and-error on wake to figure out which keyboard
it's configured for and manually switching.

Split all the keyboard related functions to its own file.