Use gem exec to run commit_format gem
Use gem exec instead of my own rolled script that checks if the gem is
installed and locks it to a specific Ruby version.

Gem exec will install it before running it in whatever Ruby version is
currently active. That way I also don't need to keep updating the Ruby
version I want it installed in.