Sleep before lock screen and sleep
Too often does my laptop unlock or wake up again after pressing these
shortcuts, while the script themselves do not. Perhaps it's the 'key up'
that wakes the machine again. Add a sleep to allow for that key up event
to be registered first before the lock or sleep script.