Improve ripgrep configuration
Configure ripgreprc to not rely on the `.gitignore` file for filtering
files in a project. A lot of times I want to still search and use FZF to
open these Git ignored files.

Update the `.ignore_global` to also contain the defaults from the
`.gitignore_global`, except some things I do want to search for.