Reorganize vim key mappings
Move more key mappings to under the leader key. Move them to two letter
mappings so they're grouped and not all over the keyboard.

Also rely a little less on the Command key, and use the leader key more
so it's easier to use the muscle memory in the terminal, where there's
no Command key available.